pyproject.toml
[tool.poetry]
name = "helena-api"
version = "0.1.0"
description = ""
authors = ["luzk <me@lucasgomide.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.8.4"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
